refactor(cairn-bench): split scorecard into subcommand for issue #99 gates
Move the top-level scorecard CLI body into crates/cairn-bench/src/scorecard/
(mod.rs exposes ScorecardArgs + run(); runner.rs holds the lifted pipeline).
Replace main.rs with a clap subcommand dispatcher routing scorecard, latency,
memory, privacy, and all — the latter four are stubs until Tasks 4-10.

Add cli_smoke integration tests asserting scorecard --help and top-level
--help list all expected subcommands.
