+++ crates/cairn-bench/benches/latency.rs
@@ -1,1 +1,2 @@
+//! Latency criterion bench targets (stub — wired in Task 4).
 fn main() {}
