+++ crates/cairn-bench/benches/lifecycle.rs
@@ -1,1 +1,2 @@
+//! Lifecycle bench targets (stub — wired in a later task).
 fn main() {}
