+++ crates/cairn-bench/src/lib.rs
@@ -1,13 +1,6 @@
 #![forbid(unsafe_code)]
 #![warn(missing_docs)]
-//! `BrainBench` retrieval scorecard runner library.
-//!
-//! Hosts the fixture loader, IR metric helpers, the adapter trait + four
-//! cairn-side implementations (`bm25-only`, `vector-bge`, `hybrid-bge-rrf`,
-//! `hybrid-openai-rrf`), and the disk-backed embedding cache that lets
-//! re-runs skip inference and HTTP cost. The report writer (Task 12)
-//! lives behind a stub today; the binary in `src/main.rs` calls every
-//! module here.
+//! Cairn bench harness: `BrainBench` scorecard + release gates (latency, memory, privacy).
 
 pub mod adapter;
 pub mod cache;
@@ -15,3 +8,4 @@ pub mod cached_embedder;
 pub mod fixture;
 pub mod metrics;
 pub mod report;
+pub mod scorecard;
